* fixed MsgGMUpdate  (float's can not be used for absolute times) * added bz.GetPlayerOpacity()
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21045 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/GameKeeper.h
@@ -50,10 +50,10 @@ public:
   ShotInfo() : present(false) {};
 
   FiringInfo firingInfo;
-  int	salt;
-  float      expireTime;
-  bool       present;
-  bool       running;
+  int    salt;
+  double expireTime;
+  bool   present;
+  bool   running;
 };
 
 class PlayerCaps
